Re-arrange files in the hardhat plugin
